% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTablesAndFigures.R
\name{createTableAndFigures}
\alias{createTableAndFigures}
\title{Create tables and figures}
\usage{
createTableAndFigures(outputFolder)
}
\arguments{
\item{outputFolder}{The path to the output folder containing the results.}
}
\description{
Create tables and figures
}
\details{
Creates tables and figures for viewing and interpreting the results. Requires that the \code{\link{execute}} function has
completed first.
}

