% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageResults.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(connectionDetails, cdmDatabaseSchema, outputFolder,
  minCellCount = 5)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included in the results.}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
